# FairSplit.Backend/src/FairSplit.Api/Presentation/Models/Responses/SettlementResponse.cs
namespace FairSplit.Api.Presentation.Models.Responses;

public sealed class SettlementResponse
{
    public Guid Id { get; set; }
    public Guid GroupId { get; set; }
    public Guid FromMemberId { get; set; }
    public Guid ToMemberId { get; set; }
    public decimal Amount { get; set; }
    public DateTimeOffset SettledAtUtc { get; set; }
}