# openapi.yaml
openapi: 3.0.3
info:
  title: FairSplit API
  version: 1.0.0
  description: API for FairSplit, a mobile group expense sharing application backend built with ASP.NET Core and PostgreSQL.
servers:
  - url: http://localhost:5000/api
    description: Local development server

tags:
  - name: Auth
    description: Token issuance for local development and demo access
  - name: Groups
    description: Group management and joining
  - name: Members
    description: Group membership read operations
  - name: Expenses
    description: Expense creation and history
  - name: Settlements
    description: Settlement recording and history
  - name: Balances
    description: Read-only derived balances

paths:
  /api/auth/dev-token:
    post:
      tags: [Auth]
      summary: Issue a development access token
      description: Returns a short-lived JWT for local demo use.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevTokenRequest'
      responses:
        '200':
          description: Access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/groups:
    get:
      tags: [Groups]
      summary: List groups
      description: Returns all groups visible in the current application scope.
      responses:
        '200':
          description: Groups retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Groups]
      summary: Create a group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/groups/{groupId}:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    get:
      tags: [Groups]
      summary: Get group details
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Group retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/groups/{groupId}/join:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    post:
      tags: [Groups]
      summary: Join group by invite code or token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinGroupRequest'
      responses:
        '200':
          description: Joined group successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/groups/{groupId}/members:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    get:
      tags: [Members]
      summary: List members in a group
      description: Returns members for a group. Current implementation does not require bearer auth.
      responses:
        '200':
          description: Members retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/groups/{groupId}/expenses:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    post:
      tags: [Expenses]
      summary: Create an expense in a group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags: [Expenses]
      summary: List expense history for a group
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Expenses retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExpenseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/groups/{groupId}/expenses/{expenseId}:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - $ref: '#/components/parameters/ExpenseId'
    get:
      tags: [Expenses]
      summary: Get details of one expense
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Expense retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/groups/{groupId}/settlements:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    post:
      tags: [Settlements]
      summary: Record a settlement in a group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSettlementRequest'
      responses:
        '201':
          description: Settlement created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags: [Settlements]
      summary: List settlements for a group
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Settlements retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SettlementResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/groups/{groupId}/settlements/{settlementId}:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - $ref: '#/components/parameters/SettlementId'
    get:
      tags: [Settlements]
      summary: Get details of one settlement
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Settlement retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/groups/{groupId}/balances:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    get:
      tags: [Balances]
      summary: Get derived balances for a group
      description: Read-only endpoint returning derived net balances per member.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Balances retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BalanceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/groups/{groupId}/balances/{memberId}:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - $ref: '#/components/parameters/MemberId'
    get:
      tags: [Balances]
      summary: Get derived balance for one member
      description: Read-only endpoint returning a single member's net balance in the group.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Balance retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    GroupId:
      in: path
      name: groupId
      required: true
      schema:
        type: string
        format: uuid
      example: 2f86f0d9-12da-44e8-a58b-8c84eec1ca87
    MemberId:
      in: path
      name: memberId
      required: true
      schema:
        type: string
        format: uuid
      example: 9f6f72f7-f59f-4c66-9ab7-393f0da34d50
    ExpenseId:
      in: path
      name: expenseId
      required: true
      schema:
        type: string
        format: uuid
      example: b1e87d98-fd8a-4a8a-95f1-35f698f2f8a8
    SettlementId:
      in: path
      name: settlementId
      required: true
      schema:
        type: string
        format: uuid
      example: 9e5d4f75-f87a-4e06-9f61-1235e74f4f9c

  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user is not allowed to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with current resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Business rule validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    CreateGroupRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Spring Trip 2026

    GroupResponse:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
          example: 2f86f0d9-12da-44e8-a58b-8c84eec1ca87
        name:
          type: string
          example: Spring Trip 2026

    JoinGroupRequest:
      type: object
      description: One of inviteCode or inviteToken should be provided.
      properties:
        inviteCode:
          type: string
          example: TRIP26
        inviteToken:
          type: string
          example: 2A7M-FAIR-SPLIT-TOKEN
        displayName:
          type: string
          example: Aigerim

    MemberResponse:
      type: object
      required: [id, groupId, displayName]
      properties:
        id:
          type: string
          format: uuid
          example: 9f6f72f7-f59f-4c66-9ab7-393f0da34d50
        groupId:
          type: string
          format: uuid
          example: 2f86f0d9-12da-44e8-a58b-8c84eec1ca87
        displayName:
          type: string
          example: Aigerim

    CreateExpenseRequest:
      type: object
      required: [payerMemberId, amount, splitType, participants]
      properties:
        payerMemberId:
          type: string
          format: uuid
          example: 4f09d2e8-7401-4bcb-9a81-6507df6c3abd
        amount:
          type: number
          format: decimal
          minimum: 0.01
          example: 120.50
        splitType:
          type: string
          enum: [equal, custom]
          example: equal
        participants:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateExpenseParticipantShare'

    CreateExpenseParticipantShare:
      type: object
      required: [memberId]
      properties:
        memberId:
          type: string
          format: uuid
          example: 9f6f72f7-f59f-4c66-9ab7-393f0da34d50
        shareAmount:
          type: number
          format: decimal
          nullable: true
          description: Required for custom split, ignored for equal split.
          example: 60.25

    ExpenseResponse:
      type: object
      required: [id, groupId, payerMemberId, amount, createdAtUtc]
      properties:
        id:
          type: string
          format: uuid
          example: b1e87d98-fd8a-4a8a-95f1-35f698f2f8a8
        groupId:
          type: string
          format: uuid
          example: 2f86f0d9-12da-44e8-a58b-8c84eec1ca87
        payerMemberId:
          type: string
          format: uuid
          example: 4f09d2e8-7401-4bcb-9a81-6507df6c3abd
        amount:
          type: number
          format: decimal
          example: 120.50
        createdAtUtc:
          type: string
          format: date-time
          example: '2026-04-18T12:30:00Z'

    ExpenseParticipantResponse:
      type: object
      required: [memberId, shareAmount]
      properties:
        memberId:
          type: string
          format: uuid
          example: 9f6f72f7-f59f-4c66-9ab7-393f0da34d50
        shareAmount:
          type: number
          format: decimal
          example: 60.25

    ExpenseDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ExpenseResponse'
        - type: object
          required: [participants]
          properties:
            participants:
              type: array
              items:
                $ref: '#/components/schemas/ExpenseParticipantResponse'

    CreateSettlementRequest:
      type: object
      required: [fromMemberId, toMemberId, amount]
      properties:
        fromMemberId:
          type: string
          format: uuid
          example: da2f14be-c29a-45de-b4cb-7f28e6119a5d
        toMemberId:
          type: string
          format: uuid
          example: 4f09d2e8-7401-4bcb-9a81-6507df6c3abd
        amount:
          type: number
          format: decimal
          minimum: 0.01
          example: 40.00

    SettlementResponse:
      type: object
      required: [id, groupId, fromMemberId, toMemberId, amount, settledAtUtc]
      properties:
        id:
          type: string
          format: uuid
          example: 9e5d4f75-f87a-4e06-9f61-1235e74f4f9c
        groupId:
          type: string
          format: uuid
          example: 2f86f0d9-12da-44e8-a58b-8c84eec1ca87
        fromMemberId:
          type: string
          format: uuid
          example: da2f14be-c29a-45de-b4cb-7f28e6119a5d
        toMemberId:
          type: string
          format: uuid
          example: 4f09d2e8-7401-4bcb-9a81-6507df6c3abd
        amount:
          type: number
          format: decimal
          example: 40.00
        settledAtUtc:
          type: string
          format: date-time
          example: '2026-04-18T15:00:00Z'

    BalanceResponse:
      type: object
      required: [groupId, memberId, netAmount]
      properties:
        groupId:
          type: string
          format: uuid
          example: 2f86f0d9-12da-44e8-a58b-8c84eec1ca87
        memberId:
          type: string
          format: uuid
          example: 9f6f72f7-f59f-4c66-9ab7-393f0da34d50
        netAmount:
          type: number
          format: decimal
          description: Positive means the member is owed money; negative means the member owes money.
          example: -35.25

    ErrorResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
          example: validation_error
        message:
          type: string
          example: splitType must be 'equal' or 'custom'
        traceId:
          type: string
          example: 00-7f4dbf53f4da8f4393037a0ad5b7c41d-4ba6ce80f94d9a9d-00

    DevTokenRequest:
      type: object
      required: [memberId, displayName]
      properties:
        memberId:
          type: string
          format: uuid
          example: 22222222-2222-2222-2222-222222222222
        displayName:
          type: string
          example: Alice

    AuthTokenResponse:
      type: object
      required: [accessToken, tokenType, expiresAtUtc, memberId, displayName]
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        tokenType:
          type: string
          example: Bearer
        expiresAtUtc:
          type: string
          format: date-time
          example: '2026-05-10T18:00:00Z'
        memberId:
          type: string
          format: uuid
          example: 22222222-2222-2222-2222-222222222222
        displayName:
          type: string
          example: Alice